requirements.txt
requests
Pillow
selenium
pyTelegramBotAPI
